Move buffered output to its own file
