+++ exo/orchestration/node.py
@@ -4,12 +4,13 @@ import asyncio
 import uuid
 import time
 import traceback
-from typing import List, Dict, Optional, Tuple, Union, Set
+from typing import List, Dict, Optional, Tuple
+from exo.inference.buffered_output import BufferedOutput
 from exo.networking import Discovery, PeerHandle, Server
-from exo.inference.inference_engine import InferenceEngine, Shard
+from exo.inference.inference_engine import Shard
 from exo.topology.topology import Topology
 from exo.topology.device_capabilities import device_capabilities, UNKNOWN_DEVICE_CAPABILITIES
-from exo.topology.partitioning_strategy import Partition, PartitioningStrategy, map_partitions_to_shards
+from exo.topology.partitioning_strategy import PartitioningStrategy, map_partitions_to_shards
 from exo.inference.generation_options import GenerationOptions
 from exo import DEBUG
 from exo.helpers import AsyncCallbackSystem
@@ -17,108 +18,6 @@ from exo.viz.topology_viz import TopologyViz
 from exo.download.download_progress import RepoProgressEvent
 from exo.inference.inference_engine import get_inference_engine, InferenceEngine
 from exo.download.shard_download import ShardDownloader
-
-class BufferedOutput:
-  stop_sequences: List[str]
-  max_tokens: int
-  eos_token_id: int
-  buffer_char_size: int
-
-  _token_count: int = 0
-  buffer: List[Tuple[int, str]]
-
-  is_finished: bool = False
-  finish_reason: Optional[str] = None
-
-  def __init__(self, max_tokens: int, eos_token_id: int, stop_sequences: List[str], tokenizer):
-    self.buffer = []
-    self.buffer_char_size = max(len(stop_sequence) for stop_sequence in stop_sequences) if len(stop_sequences) > 0 else 0
-    self.max_tokens = max_tokens
-    self.eos_token_id = eos_token_id
-    self.stop_sequences = stop_sequences
-    self.tokenizer = tokenizer
-
-  def append(self, token: int):
-    self.buffer.append((token, self.tokenizer.decode([token])))
-    self._token_count += 1
-
-    if token == self.eos_token_id:
-      self.is_finished = True
-      self.finish_reason = "stop"
-    elif self._token_count >= self.max_tokens:
-      self.is_finished = True
-      self.finish_reason = "length"
-    elif len(self.stop_sequences) > 0:
-      self.attempt_to_match_stop_sequences()
-
-  def assembled_text(self) -> str:
-    return "".join([text for _, text in self.buffer])
-
-  def attempt_to_match_stop_sequences(self):
-    assembled_text = self.assembled_text()
-    if DEBUG >= 2: print(f"Attempting to match stop sequences against: {assembled_text=}")
-
-    for stop_sequence in self.stop_sequences:
-      if len(assembled_text) < len(stop_sequence):
-        continue
-
-      if DEBUG >= 2: print(f"Checking if {assembled_text=} matches {stop_sequence=}")
-
-      if stop_sequence in assembled_text:
-        if DEBUG >= 2: print(f"Match found: {assembled_text=} matches {stop_sequence=}")
-
-        # Find character index where stop sequence starts
-        char_idx = assembled_text.index(stop_sequence)
-
-        # Find which token contains this character index and where in that token the sequence starts
-        current_char_pos = 0
-        tokens_to_keep = 0
-        for _, text in self.buffer:
-          next_char_pos = current_char_pos + len(text)
-          if current_char_pos <= char_idx < next_char_pos:
-            # Found the token containing the stop sequence
-            # Get the text before the stop sequence
-            token_offset = char_idx - current_char_pos
-            truncated_text = text[:token_offset]
-
-            # This is a little bit of a hack as the SendResults GRPC call expects tokens so to return the truncated text
-            # we need to retokenize it. This is not ideal as it means we are not returning the exact tokens that were
-            # generated by the model. However, it is the simplest way to handle this case.
-
-            # Retokenize the truncated text
-            new_tokens = self.tokenizer.encode(truncated_text, add_special_tokens=False)
-
-            # Replace the final token with the retokenized truncated text
-            self.buffer = self.buffer[:tokens_to_keep]
-            for token in new_tokens:
-              self.buffer.append((token, self.tokenizer.decode([token])))
-            break
-
-          current_char_pos = next_char_pos
-          tokens_to_keep += 1
-        else:
-          # If we didn't find the token, just keep everything up to char_idx
-          self.buffer = self.buffer[:tokens_to_keep]
-
-        self.is_finished = True
-        self.finish_reason = "stop"
-        break
-
-  def token_count(self) -> int:
-    return self._token_count
-
-  def next_tokens(self) -> List[int]:
-    if self.is_finished:
-      # Return all remaining tokens if finished
-      tokens = [token for token, _ in self.buffer]
-      self.buffer = []
-      return tokens
-    elif len(self.assembled_text()) >= self.buffer_char_size:
-      token, _ = self.buffer.pop(0)
-      return [token]
-
-    # Not enough tokens yet
-    return []
 
 
 class Node:
